Change compareWithTitle logic
Change the behaviour of the compareWithTitle option such that it is
treated the same as any of the ignore* options, namely as an additional
matching rule that contrains the results of the duplicate search even
further.

More precisely, now if option.compareWithTitle===true, then tabs with
the same URL (subject to the other matching rules) only match iff they
also share the same title.
Previously, matching in URLs alone would have sufficed.